+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat(sdk): add `get_datasample` method (#171)
 - feat: add the get_performances(compute_plan.key) function to the client (#156)
+- BREAKING CHANGE: the user must give the key of the compute plan when creating a new compute plan (#173)
 
 ### Changed